Add 64-bit ABI and debugger protocol support
This change updates the compiler and runtime to native 64-bit pointer and variable-space handling, including wider ref/index tracking, 64-bit ASM operand generation, and versioned debugger variable-table/value serialization. It also includes runtime-memory safety fixes, crash diagnostics, package bootstrap path fallback handling, and compatibility updates for pointer-sized DBP commands and file APIs.

+++ DBProCompiler/DBPCompiler/ReferenceTracker.h
@@ -70,7 +70,7 @@ public:
     };
 
     using SymbolResolver =
-        std::function<std::optional<std::uint32_t>(const ParsedReference&)>;
+        std::function<std::optional<std::uint64_t>(const ParsedReference&)>;
 
     CReferenceTracker();
     ~CReferenceTracker() = default;
